+++ app/controllers/chapters_controller.rb
@@ -21,12 +21,12 @@ class ChaptersController < ApplicationController
   # GET /chapters/1
   def show
     @page = (params[:page].to_i > 1 ? params[:page].to_i : 1)
-    reply_scope = @chapter.replies.includes(:chapter).where(reply_id: nil).page(@page).per(Chapter::REPLIES_PER_PAGE)
+    reply_scope = @chapter.replies.includes(:chapter).where(reply_id: nil).page(@page).per(Reply::REPLIES_PER_PAGE)
     last_reply_id = reply_scope.last.id
     reply_scope = reply_scope.shadow_banned(current_user ? current_user.id : nil) unless current_user && current_user.owner?
     @replies = reply_scope
     @chapter.increment! :view_count
-    current_user.read_chapter!(@chapter, last_reply_id) if current_user
+    current_user.read_parent!(@chapter, last_reply_id) if current_user
   end
 
   # GET /chapters/new