+++ db/migrate/20170404173902_add_last_reply_at_to_broadcasts.rb
@@ -1,0 +1,6 @@
+class AddLastReplyAtToBroadcasts < ActiveRecord::Migration[5.0]
+  def change
+    add_column :broadcasts, :last_reply_at, :datetime
+    add_index :broadcasts, :last_reply_at
+  end
+end