+++ test/controllers/async/blog_controller_test.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require 'test_helper'
-
-class Async::BlogControllerTest < ActionController::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end