+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### Refactoring
 - Removed typeahead library
+- Combined blog post replies and forum topic replies into one model
 
 ## 15.0.2 (January 23, 2017)